改级联 还需要测试

+++ src/main/java/cn/moquan/service/ClassGradeService.java
@@ -1,10 +1,6 @@
 package cn.moquan.service;
 
-import cn.moquan.bean.ClassGradeTeacher;
-import cn.moquan.bean.Classroom;
-import cn.moquan.bean.Student;
-import cn.moquan.bean.TeachCourseInfo;
-import cn.moquan.bean.ClassGrade;
+import cn.moquan.bean.*;
 import cn.moquan.dao.ClassGradeDao;
 import cn.moquan.util.CommonResponseBody;
 import cn.moquan.util.RollBackException;
@@ -36,6 +32,8 @@ public class ClassGradeService {
     ClassroomService classroomService;
     @Autowired
     ClassGradeTeacherService classGradeTeacherService;
+    @Autowired
+    StudentTeacherService studentTeacherService;
 
 
     public List<ClassGrade> getClassGrade(ClassGrade classGradeInfo) {
@@ -57,8 +55,7 @@ public class ClassGradeService {
             oldClassroom.setRealId(nowClassroomRealId);
 
             Classroom newClassroom = new Classroom();
-            // 标识为新添加
-            newClassroom.setId(-1);
+
             newClassroom.setClassName(nowClassName);
             newClassroom.setGradeName(nowGradeName);
             newClassroom.setSchoolName(nowSchoolName);
@@ -68,7 +65,38 @@ public class ClassGradeService {
                     classroomService.updateClassroomCommon(newClassroom, oldClassroom),
                     "添加班级时,更新教室信息失败, 请检查!"
             );
+
+            /*
+             * 级联更新判断
+             * 如果 学生表 和 授课信息表 存在相关信息 就进行级联更新
+             * 后续取消了 授课信息表的级联更新, 授课信息表会因为其它数据的变动而删除
+             */
+
+            Student student = new Student();
+            student.setSchoolName(nowSchoolName);
+            student.setClassroomRealId(nowClassroomRealId);
+            TeachCourseInfo teachCourseInfo = new TeachCourseInfo();
+            teachCourseInfo.setSchoolName(nowSchoolName);
+            teachCourseInfo.setClassroomRealId(nowClassroomRealId);
+
+            List<Student> studentList = studentService.getStudent(student);
+            List<TeachCourseInfo> teachCourseInfoList = teachCourseInfoService.getTeachCourseInfo(teachCourseInfo);
+            if (studentList.size() > 0 || teachCourseInfoList.size() > 0) {
+                student.setClassName(nowClassName);
+                student.setGradeName(nowGradeName);
+//                teachCourseInfo.setClassName(nowClassName);
+//                teachCourseInfo.setGradeName(nowGradeName);
+                for (Student s : studentList) {
+//                    studentService.updateStudentCommon(student, s);
+                    // 添加学生的班级信息
+                    studentService.updateStudentCommonNotCascade(student, s);
+                }
+//                for(TeachCourseInfo t : teachCourseInfoList){
+//                    teachCourseInfoService.updateClassGrade(teachCourseInfo, t);
+//                }
+            }
         }
+
 
         return classGradeDao.insertClassGrade(classGradeInfoList);
     }
@@ -103,6 +131,7 @@ public class ClassGradeService {
     public CommonResponseBody updateClassGrade(ClassGrade classGradeInfo, List<Integer> idList) {
 
         System.out.println(classGradeInfo);
+        System.out.println(idList);
 
         for (int id : idList) {
             updateOtherInfo(id, classGradeInfo);
@@ -132,14 +161,17 @@ public class ClassGradeService {
 
         List<ClassGrade> classGrade = getClassGrade(info);
 
-        for (ClassGrade c : classGrade) {
-            clearClassGradeInfo(c.getId());
+        if (classGrade.size() > 0) {
+            for (ClassGrade c : classGrade) {
+                clearClassGradeInfo(c.getId());
+            }
+
+            ThrowExceptionUtil.throwRollBackException(
+                    classGradeDao.deleteClassGrade(info),
+                    "删除班级信息失败, 请检查!"
+            );
         }
 
-        ThrowExceptionUtil.throwRollBackException(
-                classGradeDao.deleteClassGrade(info),
-                "删除班级信息失败, 请检查!"
-        );
         return true;
     }
 
@@ -149,7 +181,7 @@ public class ClassGradeService {
         classGrade.setId(id);
         List<ClassGrade> classGradeList = getClassGrade(classGrade);
         // 检查该班级是否存在
-        if(classGradeList.size() == 0){
+        if (classGradeList.size() == 0) {
             throw new RollBackException("该班级不存在, 请检查!",
                     new CommonResponseBody(StatusNumber.FAILED, "该班级不存在, 请检查!"));
         }
@@ -180,9 +212,10 @@ public class ClassGradeService {
         targetStudentInfo.setGradeName(gradeName);
         targetStudentInfo.setClassName(className);
         targetStudentInfo.setClassroomRealId(classroomRealId);
-        // 更新学生的班级信息
+        // 更新学生的班级信息  删除教师时 学生和老师就没有关联了
+//        studentService.updateStudentCommonNotCascade(newStudentInfo, targetStudentInfo);
         studentService.updateStudentCommon(newStudentInfo, targetStudentInfo);
-
+/*
         // 更新授课信息
         TeachCourseInfo newTeachCourseInfo = new TeachCourseInfo();
         newTeachCourseInfo.setGradeName("");
@@ -193,7 +226,16 @@ public class ClassGradeService {
         targetTeachCourseInfo.setClassName(className);
         targetTeachCourseInfo.setClassroomRealId(classroomRealId);
         // 更新
-        teachCourseInfoService.updateClassGrade(newTeachCourseInfo, targetTeachCourseInfo);
+        teachCourseInfoService.updateClassGrade(newTeachCourseInfo, targetTeachCourseInfo);*/
+
+        // 删除授课信息
+        TeachCourseInfo targetTeachCourseInfo = new TeachCourseInfo();
+        targetTeachCourseInfo.setSchoolName(schoolName);
+        targetTeachCourseInfo.setGradeName(gradeName);
+        targetTeachCourseInfo.setClassName(className);
+        targetTeachCourseInfo.setClassroomRealId(classroomRealId);
+        // 更新
+        teachCourseInfoService.deleteTeachCourseUseInfo(targetTeachCourseInfo);
 
         // 删除班级教师关联
         classGradeTeacherService.deleteClassGradeTeacherUseInfo(
@@ -202,7 +244,8 @@ public class ClassGradeService {
 
     /**
      * 班级级联更新接口
-     * @param id 更新目标的 id
+     *
+     * @param id             更新目标的 id
      * @param classGradeInfo 新的信息
      */
     public void updateOtherInfo(int id, ClassGrade classGradeInfo) {
@@ -221,28 +264,34 @@ public class ClassGradeService {
         String newSchoolName = classGradeInfo.getSchoolName();
 
         // 是否更新教室信息
-        if(newClassroomRealId != null && !"".equals(newClassroomRealId) &&
-                !newClassroomRealId.equals(oldClassroomRealId)){
+        if (newClassroomRealId != null && !"".equals(newClassroomRealId) &&
+                !newClassroomRealId.equals(oldClassroomRealId)) {
             // 更新教室
-
             // 将原来的教室置为空教室
             // 获取原来的教室信息
             Classroom oldClassroom = new Classroom();
             oldClassroom.setSchoolName(oldSchoolName);
             oldClassroom.setRealId(oldClassroomRealId);
+            oldClassroom.setGradeName(oldGradeName);
+            oldClassroom.setClassName(oldClassName);
+
             // 获取到信息
             List<Classroom> classroomList = classroomService.getClassroom(oldClassroom);
-            if(classroomList.size() > 0){
-                // 设置班级的教室信息用于置空更新
-                Classroom nullClassroomInfo = new Classroom();
-                nullClassroomInfo.setClassName("");
-                nullClassroomInfo.setGradeName("");
-                // 开始更新教室信息, 这个更新不需要教室进行级联更新
-                classroomService.updateClassroomNotCascading(nullClassroomInfo, oldClassroom);
-            }else{
-                // 原教室不存在 报错 信息错误
-                throw new RollBackException("更新班级教师时, 原教室查询失败, 请检查!",
-                        new CommonResponseBody(StatusNumber.FAILED, "更新班级教师时, 原教室查询失败, 请检查!"));
+
+            // 原班级为空 不需要置空
+            if (!"".equals(oldClassroomRealId)) {
+                if (classroomList.size() > 0) {
+                    // 设置班级的教室信息用于置空更新
+                    Classroom nullClassroomInfo = new Classroom();
+                    nullClassroomInfo.setClassName("");
+                    nullClassroomInfo.setGradeName("");
+                    // 开始更新教室信息, 这个更新不需要教室进行级联更新
+                    classroomService.updateClassroomNotCascading(nullClassroomInfo, oldClassroom);
+                } else {
+                    // 原教室不存在 报错 信息错误
+                    throw new RollBackException("更新班级教室时, 原教室查询失败, 请检查!",
+                            new CommonResponseBody(StatusNumber.FAILED, "更新班级教室时, 原教室查询失败, 请检查!"));
+                }
             }
 
             // 将空教室更新为当前使用
@@ -251,23 +300,24 @@ public class ClassGradeService {
             newClassroom.setRealId(newClassroomRealId);
             newClassroom.setSchoolName(newSchoolName);
             classroomList = classroomService.getClassroom(newClassroom);
-            if(classroomList.size() > 0){
+            if (classroomList.size() > 0) {
                 // 设置新教室的更新参数
                 Classroom newClassroomInfo = new Classroom();
                 newClassroomInfo.setClassName(newClassName);
                 newClassroomInfo.setGradeName(newGradeName);
                 // 开始更新新的教室号, 不需要级联更新
                 classroomService.updateClassroomNotCascading(newClassroomInfo, newClassroom);
-            }else{
+            } else {
                 // 新教室不存在 报错 信息错误
                 throw new RollBackException("更新班级教师时, 新教室查询失败, 请检查!",
                         new CommonResponseBody(StatusNumber.FAILED, "更新班级教师时, 新教室查询失败, 请检查!"));
             }
-        }else{// 更新教室号完成
+        } else {// 更新教室号完成
 
             Classroom newClassroom = new Classroom();
             newClassroom.setGradeName(newGradeName);
             newClassroom.setClassName(newClassName);
+            newClassroom.setSchoolName(newSchoolName);
             Classroom targetClassroom = new Classroom();
             targetClassroom.setSchoolName(oldSchoolName);
             targetClassroom.setGradeName(oldGradeName);
@@ -277,6 +327,7 @@ public class ClassGradeService {
             classroomService.updateClassroomNotCascading(newClassroom, targetClassroom);
 
         }// end if 教室信息更新完成 开始更新级联信息
+
 
         // 更新学生信息表中的数据
         // 准备需要更新的数据
@@ -292,7 +343,8 @@ public class ClassGradeService {
         targetStudentInfo.setClassName(oldClassName);
         targetStudentInfo.setClassroomRealId(oldClassroomRealId);
         // 级联更新学生信息
-        studentService.updateStudentCommon(newStudentInfo, targetStudentInfo);
+//        studentService.updateStudentCommon(newStudentInfo, targetStudentInfo);
+        studentService.updateStudentCommonNotCascade(newStudentInfo, targetStudentInfo);
 
         // 更新授课信息表中的数据
         // 准备需要更新的数据
